% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R
\docType{class}
\name{ClareExperiment-class}
\alias{ClareExperiment-class}
\alias{ClareExperiment}
\alias{ClareExperimentFromZscores}
\title{An S4 class to represent a longitudinal highly-multiplexed serology experiment}
\usage{
ClareExperiment(se)

ClareExperimentFromZscores(zscores, colData, rowData, ...)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{zscores}{A matrix of Z scores}

\item{colData}{A data frame of sample names and data. Number of rows must match the number of columns present in `zscores`}

\item{rowData}{A data frame of peptide id's and data. Number of rows must match number of rows in `zscores`}

\item{...}{arguments provided to \code{SummarizedExperiment} including rowRanges and metadata.}
}
\value{
A ClareExperiment object
}
\description{
An S4 class to represent a longitudinal highly-multiplexed serology experiment

ClareExperiment object and constructors
}
\examples{

zscores_mat <- matrix(runif(16), nrow = 4)
sample_metadata <- data.frame(
sample_id = factor(c("sample1", "sample2", "sample3", "sample4")),
subject_id = factor(c("subject1", "subject1", "subject2", "subject2")),
collection_date = as.Date(
"2023-01-01", "2023-02-01", "2023-03-01", "20203-04-01"
)
)
peptide_metadata <- data.frame(
peptide_id = c("PV2T_005322","PV2T_013621",
"PV2T_013622", "PV2T_013623"),
species = c("speciesA", "speciesB", "speciesC", "speciesD")
)
new_expt <- ClareExperimentFromZscores(
zscores = zscores_mat,
colData = sample_metadata,
rowData = peptide_metadata
)
}
