% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PV2T_rowData.R
\docType{data}
\name{PV2T_rowData}
\alias{PV2T_rowData}
\title{PV2T PepSeq peptide library}
\format{
A data frame with 14,876 rows and 6 columns

\describe{
   \item{library_member}{peptide identifier for PepSeq assay}
   \item{species}{viral species to which the peptide belongs}
   \item{annotation}{taxon id corresponding to the viral species}
   \item{gene}{viral gene, if known}
   \item{peptide}{peptide sequence}
   \item{uniprot_id}{uniprot id for the protein}
}
}
\source{
https://osf.io/6ht43/
}
\usage{
PV2T_rowData
}
\description{
PV2T PepSeq peptide library
}
\keyword{datasets}
