% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBprog_colData.R
\docType{data}
\name{TBprog_colData}
\alias{TBprog_colData}
\title{TB progessor cohort data}
\format{
A data frame with 195 rows and 9 columns

\describe{
   \item{sample_id}{sample identifier for PepSeq assay}
   \item{subject_id}{study subject identifier}
   \item{collection_date}{date of sample collection}
   \item{visit}{approximate day of sample collection, eg 180 days from the first collection}
   \item{sex}{a subject's sex}
   \item{age_at_last_birthday}{a subject's age at their last birthday}
   \item{ethnicity}{a subject's ethnicity}
   \item{scriba2018_progressor_or_control}{control did not progress to active tuberculosis disease; progressor developed active tuberculosis disease during the study}

}
}
\source{
https://doi.org/10.1371/journal.ppat.1006687
}
\usage{
TBprog_colData
}
\description{
TB progessor cohort data
}
\keyword{datasets}
