% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBprog_zscores.R
\docType{data}
\name{TBprog_zscores}
\alias{TBprog_zscores}
\title{TBprog_zscores}
\format{
A data frame with 14876 rows and 196 columns

\describe{
     \item{Sequence name}{name of the peptide in the PepSeq library}
 \item{ S040333_D180.Pro_PV2T }{names of samples assayed}
 \item{ S040333_D360.Pro_PV2T }{names of samples assayed}
 \item{ S040333_D540.Pro_PV2T }{names of samples assayed}
 \item{ S040695_D180.Pro_PV2T }{names of samples assayed}
 \item{ S040695_D360.Pro_PV2T }{names of samples assayed}
 \item{ S040695_D540.Pro_PV2T }{names of samples assayed}
 \item{ S040699_D180.Pro_PV2T }{names of samples assayed}
 \item{ S040699_D360.Pro_PV2T }{names of samples assayed}
 \item{ S040699_D540.Pro_PV2T }{names of samples assayed}
 \item{ S040718_D180.Pro_PV2T }{names of samples assayed}
 \item{ S040718_D360.Pro_PV2T }{names of samples assayed}
 \item{ S040718_D540.Pro_PV2T }{names of samples assayed}
 \item{ S040741_D180.Pro_PV2T }{names of samples assayed}
 \item{ S040741_D360.Pro_PV2T }{names of samples assayed}
 \item{ S040741_D540.Pro_PV2T }{names of samples assayed}
 \item{ S040811_D180.Pro_PV2T }{names of samples assayed}
 \item{ S040811_D360.Pro_PV2T }{names of samples assayed}
 \item{ S040811_D540.Pro_PV2T }{names of samples assayed}
 \item{ S040843_D180.Pro_PV2T }{names of samples assayed}
 \item{ S040843_D360.Pro_PV2T }{names of samples assayed}
 \item{ S040843_D540.Pro_PV2T }{names of samples assayed}
 \item{ S040886_D180.Pro_PV2T }{names of samples assayed}
 \item{ S040886_D360.Pro_PV2T }{names of samples assayed}
 \item{ S040886_D540.Pro_PV2T }{names of samples assayed}
 \item{ S040979_D180.Pro_PV2T }{names of samples assayed}
 \item{ S040979_D360.Pro_PV2T }{names of samples assayed}
 \item{ S040979_D540.Pro_PV2T }{names of samples assayed}
 \item{ S041003_D180.Pro_PV2T }{names of samples assayed}
 \item{ S041003_D360.Pro_PV2T }{names of samples assayed}
 \item{ S041003_D540.Pro_PV2T }{names of samples assayed}
 \item{ S041020_D180.Pro_PV2T }{names of samples assayed}
 \item{ S041020_D360.Pro_PV2T }{names of samples assayed}
 \item{ S041020_D540.Pro_PV2T }{names of samples assayed}
 \item{ S041033_D180.Pro_PV2T }{names of samples assayed}
 \item{ S041033_D360.Pro_PV2T }{names of samples assayed}
 \item{ S041033_D540.Pro_PV2T }{names of samples assayed}
 \item{ S041037_D180.Pro_PV2T }{names of samples assayed}
 \item{ S041037_D360.Pro_PV2T }{names of samples assayed}
 \item{ S041037_D540.Pro_PV2T }{names of samples assayed}
 \item{ S041067_D180.Pro_PV2T }{names of samples assayed}
 \item{ S041067_D360.Pro_PV2T }{names of samples assayed}
 \item{ S041067_D540.Pro_PV2T }{names of samples assayed}
 \item{ S041104_D180.Pro_PV2T }{names of samples assayed}
 \item{ S041104_D360.Pro_PV2T }{names of samples assayed}
 \item{ S041104_D540.Pro_PV2T }{names of samples assayed}
 \item{ S041120_D180.Pro_PV2T }{names of samples assayed}
 \item{ S041120_D360.Pro_PV2T }{names of samples assayed}
 \item{ S041120_D540.Pro_PV2T }{names of samples assayed}
 \item{ S041125_D180.Pro_PV2T }{names of samples assayed}
 \item{ S041125_D360.Pro_PV2T }{names of samples assayed}
 \item{ S041125_D540.Pro_PV2T }{names of samples assayed}
 \item{ S041140_D180.Pro_PV2T }{names of samples assayed}
 \item{ S041140_D360.Pro_PV2T }{names of samples assayed}
\item{ S041140_D540.Pro_PV2T }{names of samples assayed}
\item{ S060129_D180.Pro_PV2T }{names of samples assayed}
\item{ S060129_D360.Pro_PV2T }{names of samples assayed}
\item{ S060129_D540.Pro_PV2T }{names of samples assayed}
\item{ S060231_D180.Pro_PV2T }{names of samples assayed}
\item{ S060231_D360.Pro_PV2T }{names of samples assayed}
\item{ S060231_D540.Pro_PV2T }{names of samples assayed}
\item{ S070056_D180.Pro_PV2T }{names of samples assayed}
\item{ S070056_D360.Pro_PV2T }{names of samples assayed}
\item{ S070056_D540.Pro_PV2T }{names of samples assayed}
\item{ S070074_D180.Pro_PV2T }{names of samples assayed}
\item{ S070074_D360.Pro_PV2T }{names of samples assayed}
\item{ S070074_D540.Pro_PV2T }{names of samples assayed}
\item{ S070119_D180.Pro_PV2T }{names of samples assayed}
\item{ S070119_D360.Pro_PV2T }{names of samples assayed}
\item{ S070119_D540.Pro_PV2T }{names of samples assayed}
\item{ S070140_D180.Pro_PV2T }{names of samples assayed}
\item{ S070140_D360.Pro_PV2T }{names of samples assayed}
\item{ S070140_D540.Pro_PV2T }{names of samples assayed}
\item{ S070157_D180.Pro_PV2T }{names of samples assayed}
\item{ S070157_D360.Pro_PV2T }{names of samples assayed}
\item{ S070157_D540.Pro_PV2T }{names of samples assayed}
\item{ S070167_D180.Pro_PV2T }{names of samples assayed}
\item{ S070167_D360.Pro_PV2T }{names of samples assayed}
\item{ S070167_D540.Pro_PV2T }{names of samples assayed}
\item{ S070177_D180.Pro_PV2T }{names of samples assayed}
\item{ S070177_D360.Pro_PV2T }{names of samples assayed}
\item{ S070177_D540.Pro_PV2T }{names of samples assayed}
\item{ S070180_D180.Pro_PV2T }{names of samples assayed}
\item{ S070180_D360.Pro_PV2T }{names of samples assayed}
\item{ S070180_D540.Pro_PV2T }{names of samples assayed}
\item{ S070200_D180.Pro_PV2T }{names of samples assayed}
\item{ S070200_D360.Pro_PV2T }{names of samples assayed}
\item{ S070200_D540.Pro_PV2T }{names of samples assayed}
\item{ S070202_D180.Pro_PV2T }{names of samples assayed}
\item{ S070202_D360.Pro_PV2T }{names of samples assayed}
\item{ S070202_D540.Pro_PV2T }{names of samples assayed}
\item{ S070203_D180.Pro_PV2T }{names of samples assayed}
\item{ S070203_D360.Pro_PV2T }{names of samples assayed}
\item{ S070203_D540.Pro_PV2T }{names of samples assayed}
\item{ S070214_D180.Pro_PV2T }{names of samples assayed}
\item{ S070214_D360.Pro_PV2T }{names of samples assayed}
\item{ S070214_D540.Pro_PV2T }{names of samples assayed}
\item{ S070218_D180.Pro_PV2T }{names of samples assayed}
\item{ S070218_D360.Pro_PV2T }{names of samples assayed}
\item{ S070218_D540.Pro_PV2T }{names of samples assayed}
\item{ S070229_D180.Pro_PV2T }{names of samples assayed}
\item{ S070229_D360.Pro_PV2T }{names of samples assayed}
\item{ S070229_D540.Pro_PV2T }{names of samples assayed}
\item{ S070236_D180.Pro_PV2T }{names of samples assayed}
\item{ S070236_D360.Pro_PV2T }{names of samples assayed}
\item{ S070236_D540.Pro_PV2T }{names of samples assayed}
\item{ S070240_D180.Pro_PV2T }{names of samples assayed}
\item{ S070240_D360.Pro_PV2T }{names of samples assayed}
\item{ S070240_D540.Pro_PV2T }{names of samples assayed}
\item{ S070259_D180.Pro_PV2T }{names of samples assayed}
\item{ S070259_D360.Pro_PV2T }{names of samples assayed}
\item{ S070259_D540.Pro_PV2T }{names of samples assayed}
\item{ S070262_D180.Pro_PV2T }{names of samples assayed}
\item{ S070262_D360.Pro_PV2T }{names of samples assayed}
\item{ S070262_D540.Pro_PV2T }{names of samples assayed}
\item{ S070276_D180.Pro_PV2T }{names of samples assayed}
\item{ S070276_D360.Pro_PV2T }{names of samples assayed}
\item{ S070276_D540.Pro_PV2T }{names of samples assayed}
\item{ S070282_D180.Pro_PV2T }{names of samples assayed}
\item{ S070282_D360.Pro_PV2T }{names of samples assayed}
\item{ S070282_D540.Pro_PV2T }{names of samples assayed}
\item{ S070290_D180.Pro_PV2T }{names of samples assayed}
\item{ S070290_D360.Pro_PV2T }{names of samples assayed}
\item{ S070290_D540.Pro_PV2T }{names of samples assayed}
\item{ S070293_D180.Pro_PV2T }{names of samples assayed}
\item{ S070293_D360.Pro_PV2T }{names of samples assayed}
\item{ S070293_D540.Pro_PV2T }{names of samples assayed}
\item{ S070308_D180.Pro_PV2T }{names of samples assayed}
\item{ S070308_D360.Pro_PV2T }{names of samples assayed}
\item{ S070308_D540.Pro_PV2T }{names of samples assayed}
\item{ S070313_D180.Pro_PV2T }{names of samples assayed}
\item{ S070313_D360.Pro_PV2T }{names of samples assayed}
\item{ S070313_D540.Pro_PV2T }{names of samples assayed}
\item{ S070355_D180.Pro_PV2T }{names of samples assayed}
\item{ S070355_D360.Pro_PV2T }{names of samples assayed}
\item{ S070355_D540.Pro_PV2T }{names of samples assayed}
\item{ S070360_D180.Pro_PV2T }{names of samples assayed}
\item{ S070360_D360.Pro_PV2T }{names of samples assayed}
\item{ S070360_D540.Pro_PV2T }{names of samples assayed}
\item{ S070361_D180.Pro_PV2T }{names of samples assayed}
\item{ S070361_D360.Pro_PV2T }{names of samples assayed}
\item{ S070361_D540.Pro_PV2T }{names of samples assayed}
\item{ S070369_D180.Pro_PV2T }{names of samples assayed}
\item{ S070369_D360.Pro_PV2T }{names of samples assayed}
\item{ S070369_D540.Pro_PV2T }{names of samples assayed}
\item{ S070388_D180.Pro_PV2T }{names of samples assayed}
\item{ S070388_D360.Pro_PV2T }{names of samples assayed}
\item{ S070388_D540.Pro_PV2T }{names of samples assayed}
\item{ S070390_D180.Pro_PV2T }{names of samples assayed}
\item{ S070390_D360.Pro_PV2T }{names of samples assayed}
\item{ S070390_D540.Pro_PV2T }{names of samples assayed}
\item{ S070391_D180.Pro_PV2T }{names of samples assayed}
\item{ S070391_D360.Pro_PV2T }{names of samples assayed}
\item{ S070391_D540.Pro_PV2T }{names of samples assayed}
\item{ S070392_D180.Pro_PV2T }{names of samples assayed}
\item{ S070392_D360.Pro_PV2T }{names of samples assayed}
\item{ S070392_D540.Pro_PV2T }{names of samples assayed}
\item{ S070407_D180.Pro_PV2T }{names of samples assayed}
\item{ S070407_D360.Pro_PV2T }{names of samples assayed}
\item{ S070407_D540.Pro_PV2T }{names of samples assayed}
\item{ S070424_D180.Pro_PV2T }{names of samples assayed}
\item{ S070424_D360.Pro_PV2T }{names of samples assayed}
\item{ S070424_D540.Pro_PV2T }{names of samples assayed}
\item{ S070438_D180.Pro_PV2T }{names of samples assayed}
\item{ S070438_D360.Pro_PV2T }{names of samples assayed}
\item{ S070438_D540.Pro_PV2T }{names of samples assayed}
\item{ S070440_D180.Pro_PV2T }{names of samples assayed}
\item{ S070440_D360.Pro_PV2T }{names of samples assayed}
\item{ S070440_D540.Pro_PV2T }{names of samples assayed}
\item{ S070630_D180.Pro_PV2T }{names of samples assayed}
\item{ S070630_D360.Pro_PV2T }{names of samples assayed}
\item{ S070630_D540.Pro_PV2T }{names of samples assayed}
\item{ S070663_D180.Pro_PV2T }{names of samples assayed}
\item{ S070663_D360.Pro_PV2T }{names of samples assayed}
\item{ S070663_D540.Pro_PV2T }{names of samples assayed}
\item{ S071112_D180.Pro_PV2T }{names of samples assayed}
\item{ S071112_D360.Pro_PV2T }{names of samples assayed}
\item{ S071112_D540.Pro_PV2T }{names of samples assayed}
\item{ S071115_D180.Pro_PV2T }{names of samples assayed}
\item{ S071115_D360.Pro_PV2T }{names of samples assayed}
\item{ S071115_D540.Pro_PV2T }{names of samples assayed}
\item{ S071132_D180.Pro_PV2T }{names of samples assayed}
\item{ S071132_D360.Pro_PV2T }{names of samples assayed}
\item{ S071132_D540.Pro_PV2T }{names of samples assayed}
\item{ S071135_D180.Pro_PV2T }{names of samples assayed}
\item{ S071135_D360.Pro_PV2T }{names of samples assayed}
\item{ S071135_D540.Pro_PV2T }{names of samples assayed}
\item{ S071147_D180.Pro_PV2T }{names of samples assayed}
\item{ S071147_D360.Pro_PV2T }{names of samples assayed}
\item{ S071147_D540.Pro_PV2T }{names of samples assayed}
\item{ S071153_D180.Pro_PV2T }{names of samples assayed}
\item{ S071153_D360.Pro_PV2T }{names of samples assayed}
\item{ S071153_D540.Pro_PV2T }{names of samples assayed}
\item{ S071158_D180.Pro_PV2T }{names of samples assayed}
\item{ S071158_D360.Pro_PV2T }{names of samples assayed}
\item{ S071158_D540.Pro_PV2T }{names of samples assayed}
}
}
\source{
Kelley, E. J. et al. Virome-wide detection of natural infection events and the associated antibody dynamics using longitudinal highly-multiplexed serology. Nat. Commun. 24 (In revision).
}
\usage{
data("TBprog_zscores")
}
\description{
A data set of PepSeq Z scores for the TB progessor cohort.
}
\examples{
data("TBprog_zscores")

}
\keyword{datasets}
