% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformZscores.R
\name{transform_zscores}
\alias{transform_zscores}
\title{Description of the `transform_zscores` helper function}
\usage{
transform_zscores(zscores, base = 2, offset = 3)
}
\arguments{
\item{zscores}{The input data to be transformed}

\item{base}{The base to use for the logarithmic transformation (default: 2)}

\item{offset}{The offset to use for the logarithmic transformation (default: 3)}
}
\value{
The transformed data
}
\description{
Description of the `transform_zscores` helper function
}
