% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformZscores.R
\name{transformZscores,ClareExperiment-method}
\alias{transformZscores,ClareExperiment-method}
\title{Description of the `transformZscores` S4 method}
\usage{
\S4method{transformZscores}{ClareExperiment}(object, base = 2, offset = 3, ...)
}
\arguments{
\item{object}{The input ClareExperiment object}

\item{base}{The base to use for the logarithmic transformation (default: 2)}

\item{offset}{The offset to use for the logarithmic transformation (default: 3)}

\item{...}{other arguments passed}
}
\value{
The transformed ClareExperiment object
}
\description{
Description of the `transformZscores` S4 method
}
\examples{
zscores_mat <- matrix(runif(16), nrow = 4)
sample_metadata <- data.frame(
sample_id = factor(c("sample1", "sample2", "sample3", "sample4")),
subject_id = factor(c("subject1", "subject1", "subject2", "subject2")),
collection_date = as.Date(
"2023-01-01", "2023-02-01", "2023-03-01", "20203-04-01"
)
)
peptide_metadata <- data.frame(
peptide_id = c("PV2T_005322","PV2T_013621",
"PV2T_013622", "PV2T_013623"),
species = c("speciesA", "speciesB", "speciesC", "speciesD")
)
clare_expt <- ClareExperimentFromZscores(
zscores = zscores_mat,
colData = sample_metadata,
rowData = peptide_metadata
)
se_trans <- transformZscores(clare_expt)
}
