% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{transformZscores}
\alias{transformZscores}
\title{Transform Z-scores}
\usage{
transformZscores(object, ...)
}
\arguments{
\item{object}{a Summarized Experiment object}

\item{...}{other arguments passed to transformZscores}
}
\description{
Transform Z-scores
}
